backend/requirements.txt
Django==5.0.1
djangorestframework==3.15.1
mongoengine==0.28.2
python-dotenv==1.0.1
django-cors-headers==4.3.1
